♻️ Arrumar divisão de objetos entre CronApi e ApiConsumer
