+++ backend/src/main/java/com/group/backend/controller/CronController.java
@@ -4,14 +4,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.group.backend.cron.Cron;
+import com.group.backend.cron.CronApi;
 
 @RestController
 @RequestMapping("cron")
 public class CronController {
-	public final Cron cron;
+	public final CronApi cron;
 
-	public CronController(Cron cron) {
+	public CronController(CronApi cron) {
 		this.cron = cron;
 	}
